third_party/blink/web_tests/fast/canvas-api/canvas-overloads-fillText.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
var ctx = document.createElement('canvas').getContext('2d');

test(function(t) {
    assert_throws_js(TypeError, function() {
      ctx.fillText();
    });
    assert_throws_js(TypeError, function() {
      ctx.fillText('moo');
    });
    assert_throws_js(TypeError, function() {
      ctx.fillText('moo', 0);
    });
    assert_equals(ctx.fillText('moo',0,0), undefined);
    assert_equals(ctx.fillText('moo',0,0,0), undefined);
    assert_equals(ctx.fillText('moo',0,0,0,0), undefined);
}, "Test the behavior of CanvasRenderingContext2D.fillText() when called with different numbers of arguments.");
</script>
</body>
